+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Modal to display full sized review images
+
+### Fixed
+
+- Review image thumbnail sizing
+
 ## [1.8.3] - 2021-01-08
 
 ### Added